backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
langgraph==0.2.60
langchain-google-genai==2.0.8
google-generativeai==0.8.4
pydantic==2.10.4
pydantic-settings==2.7.1
pyyaml==6.0.2
typer==0.15.1
rich==13.9.4
python-multipart==0.0.20
aiosqlite==0.20.0
tenacity==9.0.0
